Add app background (#29)
* feat: add background svg to app and set container accordingly

* refactor: delete unused Navbar and relocate ConnectWallet

* refactor: Set hydra background to cover whole screen, adapt correct ConnectWallet
