+++ src/common/components/Layout.tsx
@@ -1,23 +1,25 @@
 import { ReactNode } from "react";
 import styled, { ThemeProvider } from "styled-components";
-import { StakenetGlobalStyle } from "../../shell/theme/stakenetGlobalStyle";
-import Navbar from "./Navbar/Navbar";
 import Web3Wrapper from "./Web3Wrapper";
+import { StakenetGlobalStyle } from "../../shell/theme/stakenetGlobalStyle";
+
 import { stakenetTheme } from "../../shell/theme/stakenetTheme";
 
 interface ILayoutProps {
   theme: any;
   children: ReactNode;
 }
-
-const Root = styled.div``;
+const Root = styled.div`
+  display: flex;
+  flex-direction: column;
+  min-height: 100vh;
+`;
 
 const Layout = ({ theme, children }: ILayoutProps) => {
   return (
     <ThemeProvider theme={{ ...theme, sTheme: stakenetTheme }}>
       <StakenetGlobalStyle />
       <Web3Wrapper>
-        <Navbar />
         <Root>{children}</Root>
       </Web3Wrapper>
     </ThemeProvider>
